policy: remove bdd governance fixtures facade
