+++ .github/workflows/microcrate-ci.yml
@@ -156,7 +156,6 @@ jobs:
             adze-runtime-governance-matrix
             adze-runtime2-governance
             adze-bdd-governance-core
-            adze-bdd-governance-fixtures
             adze-bdd-grammar-fixtures
             adze-bdd-grid-core
             adze-bdd-scenario-fixtures
@@ -254,7 +253,6 @@ jobs:
         run: |
           cargo test \
             -p adze-bdd-governance-core \
-            -p adze-bdd-governance-fixtures \
             -p adze-bdd-grammar-fixtures \
             -p adze-bdd-grid-core \
             -p adze-bdd-scenario-fixtures \
