+++ book/src/microcrates.md
@@ -33,7 +33,6 @@ Quality-assurance infrastructure for feature tracking and behavioral contracts:
 | `bdd-grammar-fixtures` | `crates/bdd-grammar-fixtures/` | Test fixtures and grammar-level BDD analysis |
 | `bdd-governance-core` | `crates/bdd-governance-core/` | Governance BDD snapshots/matrix composition |
 | `bdd-governance-reporting-core` | `crates/bdd-governance-reporting-core/` | Profile-aware governance report/status formatting |
-| `bdd-governance-fixtures` | `crates/bdd-governance-fixtures/` | Governance test fixtures |
 | `bdd-scenario-fixtures` | `crates/bdd-scenario-fixtures/` | Shared BDD scenario fixtures |
 | `bdd-grid-contract` | `crates/bdd-grid-contract/` | Grid/matrix BDD contracts |
 | `bdd-grid-core` | `crates/bdd-grid-core/` | Grid BDD evaluation logic |
