+++ crates/README.md
@@ -26,7 +26,6 @@ Behavior-driven development infrastructure for scenario tracking and progress re
 | Crate | Purpose |
 |-------|---------|
 | [`bdd-governance-core`](bdd-governance-core) | Core implementation of BDD governance |
-| [`bdd-governance-fixtures`](bdd-governance-fixtures) | Test fixtures for BDD governance scenarios |
 | [`bdd-grammar-fixtures`](bdd-grammar-fixtures) | Test fixtures for grammar BDD scenarios |
 | [`bdd-grid-core`](bdd-grid-core) | Core implementation of BDD scenario grids |
 | [`bdd-scenario-fixtures`](bdd-scenario-fixtures) | General BDD scenario test fixtures |
