--- crates/bdd-governance-fixtures/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "adze-bdd-governance-fixtures"
-version = "0.1.0"
-edition.workspace = true
-rust-version.workspace = true
-license = "MIT OR Apache-2.0"
-description = "BDD progress/reporting helpers bound to parser feature profiles."
-repository = "https://github.com/EffortlessMetrics/adze"
-publish = ["crates.io"]
-keywords = ["bdd", "testing", "scenarios", "parser", "grammar"]
-categories = ["development-tools::testing"]
-homepage = "https://github.com/EffortlessMetrics/adze"
-documentation = "https://docs.rs/adze-bdd-governance-fixtures"
-
-[features]
-default = []
-strict_api = []
-strict_docs = []
-
-[dependencies]
-adze-bdd-governance-core = { version = "0.1.0", path = "../bdd-governance-core" }
-adze-governance-runtime-reporting = { version = "0.1.0", path = "../governance-runtime-reporting" }
-
-[dev-dependencies]
-proptest = { workspace = true }
