--- crates/bdd-governance-fixtures/README.md
@@ -1,10 +1,0 @@
-# adze-bdd-governance-fixtures
-
-BDD progress/reporting helpers bound to parser feature profiles.
-
-Part of the [adze](https://github.com/EffortlessMetrics/adze) workspace.
-
-## License
-
-Licensed under either of [Apache License, Version 2.0](../../LICENSE-APACHE)
-or [MIT License](../../LICENSE-MIT) at your option.
