+++ crates/bdd-scenario-fixtures/Cargo.toml
@@ -18,8 +18,9 @@ strict_api = []
 strict_docs = []
 
 [dependencies]
-adze-bdd-governance-fixtures = { version = "0.1.0", path = "../bdd-governance-fixtures" }
+adze-bdd-governance-core = { version = "0.1.0", path = "../bdd-governance-core" }
 adze-bdd-grammar-fixtures = { version = "0.1.0", path = "../bdd-grammar-fixtures" }
+adze-governance-runtime-reporting = { version = "0.1.0", path = "../governance-runtime-reporting" }
 
 [dev-dependencies]
 adze-ir = { version = "0.8.0-dev", path = "../../ir" }
