+++ docs/status/MICROCRATE_TEST_COVERAGE.md
@@ -2,25 +2,24 @@
 
 **Generated:** 2026-03-26
 **Last Updated:** 2026-03-27
-**Total Crates:** 18
+**Total Crates:** 17
 
 ## Summary
 
 | Category | Count | Percentage |
 |----------|-------|------------|
-| Complete (BDD + Property) | 18 | 100% |
+| Complete (BDD + Property) | 17 | 100% |
 | Contract Lock Tests | 30+ | 95%+ |
 
-All 18 remaining tracked microcrates have comprehensive test coverage with both BDD tests and property-based tests.
+All 17 remaining tracked microcrates have comprehensive test coverage with both BDD tests and property-based tests.
 
 ## Complete Coverage (BDD + Property Tests)
 
-All 18 remaining tracked crates have both BDD tests and property-based tests:
+All 17 remaining tracked crates have both BDD tests and property-based tests:
 
 | Crate | BDD File | Property File | Contract Lock |
 |-------|----------|---------------|---------------|
 | `bdd-governance-core` | ✓ | ✓ | ✓ |
-| `bdd-governance-fixtures` | ✓ | ✓ | - |
 | `bdd-grammar-fixtures` | ✓ | ✓ | - |
 | `bdd-grid-core` | ✓ | ✓ | ✓ |
 | `bdd-scenario-fixtures` | ✓ | ✓ | - |
@@ -61,7 +60,6 @@ The following 30+ remaining tracked crates have `contract_lock.rs` test files (c
 
 The following crates do not have contract lock tests (by design):
 
-- `bdd-governance-fixtures` - Test fixtures crate
 - `bdd-grammar-fixtures` - Test fixtures crate
 - `bdd-scenario-fixtures` - Test fixtures crate
 - `ts-c-harness` - FFI test harness (excluded from workspace)
