+++ policy/package-boundary.toml
@@ -37,20 +37,6 @@ last_changed = "2026-05-12"
 last_changed_pr = "#694"
 
 [[package]]
-name = "adze-bdd-governance-fixtures"
-path = "crates/bdd-governance-fixtures"
-category = "owner-module-migration-target"
-owner = "governance/bdd"
-production_use = true
-publish_intent = "retire-or-inline"
-support_tier_impact = "temporary internal surface; no stable product claim until migrated or promoted"
-ci_impact = "collapse target; keep out of default CI expansion unless routed by owner"
-migration_target = "governance/bdd"
-removal_or_promotion_condition = "Move into the SRP owner submodule or reclassify with an accepted spec/ADR before 0.9 release."
-last_changed = "2026-05-12"
-last_changed_pr = "#694"
-
-[[package]]
 name = "adze-bdd-grammar-fixtures"
 path = "crates/bdd-grammar-fixtures"
 category = "owner-module-migration-target"
